use the variant impacts table in inheritance tools.
This was originally for #764 but turns out to be a larger
change. If we use the variant_impacts table, then we could
group by transcript, rather than by gene. For --min-kindreds,
it probably makes sense to still group by gene.

Using only the variants table as before acted as a sort of
filter. With this change, we can show the same variant many
times, once for each of its transcripts. This mostly goes away
with "impact_severity != 'LOW'" but can still be a problem.

For comp-hets, I think we want to group by transcript.

There is some magic in here where I convert
  "impact_severity != 'LOW'"

to:
   "vi.impact_severity != 'LOW'"

and do this for any column in the variant impacts table.
Likewise for columns in the variants table.
This may cause rare problems but users can get around them
by specifying `vi.impact_severity` directly.